Renames, change in Versioning structure
-Removed VersionDeclareSpace.h and moved it's functionality into Versioning.h
-Versioning macros now prevent the use of the interface/version name to prevent the use of its own name to mean an interface version and not the implementation version. use the _NO_SFX version to make it 1 to 1
-Removed the "BasicType" type alias, instead making ITypes and AbstractTypes share the same name. The formerly named IType is now prefixed with "Basic" in its interface versions, while the internal version formerly named AbstractType is given IType. A similar convention will follow with the link to concrete types, and the prefix there is "Simple"
-Renamed Type to TypeInfo. Far too many hits occur when using the name Type

+++ include/Lexicon/Interfaces/ReportManager.h
@@ -1,7 +1,6 @@
 #pragma once
 
 
-#include "Lexicon/Versioning.h"
 //This is ignoring the precompiled header for some fuck all reason.
 #include "Lexicon/Interfaces/Interface.h"
 #include "Lexicon/Interfaces/InterfaceSingleton.h"
@@ -70,7 +69,7 @@ namespace LEX
 	//I would like to have an object be a handler for this.
 	//TODO: Put inside cpp
 	using thread_hash = uint64_t;
-	using LogEditorID = uint32_t;
+	using LogEditorID = uint64_t;
 
 
 	struct EditorHandle;
